Move worker build, move internal ts types
