+++ src/utilities/customTS.worker.ts
@@ -1,4 +1,5 @@
-import type { TypeScriptWorker, Diagnostic } from 'monaco/languages/typescript';
+import type { CompletionInfo } from 'typescript';
+import type { Diagnostic, TypeScriptWorker } from '../types/monaco-internal';
 
 // Appease the TSC - this special window object is read by the Custom Worker implementation of Monaco
 declare class TsWorkerOverride implements Partial<TypeScriptWorker> {}
@@ -12,14 +13,20 @@ declare global {
 // Implement whatever overrides we want!
 self.customTSWorkerFactory = tsw => {
   return class extends tsw implements Partial<TypeScriptWorker> {
+    private completion_name = 'test';
     constructor(...args: any) {
       super(...args);
     }
     async getCompletionsAtPosition(fileName: string, position: number) {
-      const completions = await super.getCompletionsAtPosition(fileName, position);
-      const funny_entry = completions.entries[0];
-      funny_entry.name = 'TEST COMPLETION';
-      completions.entries.unshift(funny_entry);
+      const completions = await (super.getCompletionsAtPosition(fileName, position) as Promise<
+        CompletionInfo | undefined
+      >);
+
+      const completion = completions.entries[0];
+      completion.name = this.completion_name;
+
+      completions.entries.push(completion);
+
       return completions;
     }
     async getSemanticDiagnostics(fileName: string): Promise<Diagnostic[]> {
@@ -38,6 +45,10 @@ self.customTSWorkerFactory = tsw => {
       const diagnostics = await super.getSyntacticDiagnostics(fileName);
 
       return diagnostics;
+    }
+
+    async setSuggestionName(name: string): Promise<void> {
+      this.completion_name = name;
     }
   };
 };
